LDEV-3638: Remove lesson and all its resources on demand. Files in content repository stay. Clean up unnecessary code, format.

+++ lams_tool_scratchie/src/java/org/lamsfoundation/lams/tool/scratchie/service/ScratchieServiceImpl.java
@@ -40,7 +40,6 @@ import java.util.Map;
 import java.util.Set;
 import java.util.SortedMap;
 import java.util.TreeSet;
-import java.util.concurrent.Callable;
 
 import org.apache.commons.lang.StringEscapeUtils;
 import org.apache.commons.lang.StringUtils;
@@ -67,7 +66,6 @@ import org.lamsfoundation.lams.tool.ToolOutputDefinition;
 import org.lamsfoundation.lams.tool.ToolSessionExportOutputData;
 import org.lamsfoundation.lams.tool.ToolSessionManager;
 import org.lamsfoundation.lams.tool.exception.DataMissingException;
-import org.lamsfoundation.lams.tool.exception.SessionDataExistsException;
 import org.lamsfoundation.lams.tool.exception.ToolException;
 import org.lamsfoundation.lams.tool.scratchie.ScratchieConstants;
 import org.lamsfoundation.lams.tool.scratchie.dao.ScratchieAnswerVisitDAO;
@@ -90,9 +88,7 @@ import org.lamsfoundation.lams.tool.scratchie.model.ScratchieSession;
 import org.lamsfoundation.lams.tool.scratchie.model.ScratchieUser;
 import org.lamsfoundation.lams.tool.scratchie.util.ScratchieAnswerComparator;
 import org.lamsfoundation.lams.tool.scratchie.util.ScratchieItemComparator;
-import org.lamsfoundation.lams.tool.scratchie.util.ScratchieSessionComparator;
 import org.lamsfoundation.lams.tool.scratchie.util.ScratchieToolContentHandler;
-import org.lamsfoundation.lams.tool.scratchie.web.action.LearningAction;
 import org.lamsfoundation.lams.tool.service.ILamsToolService;
 import org.lamsfoundation.lams.usermanagement.User;
 import org.lamsfoundation.lams.usermanagement.dto.UserDTO;
@@ -207,7 +203,7 @@ public class ScratchieServiceImpl implements IScratchieService, ToolContentManag
 
     @Override
     public void releaseItemsFromCache(Scratchie scratchie) {
-	for (ScratchieItem item : (Set<ScratchieItem>) scratchie.getScratchieItems()) {
+	for (ScratchieItem item : scratchie.getScratchieItems()) {
 	    scratchieItemDao.releaseItemFromCache(item);
 	}
     }
@@ -1770,23 +1766,29 @@ public class ScratchieServiceImpl implements IScratchieService, ToolContentManag
     }
 
     @Override
-    public void removeToolContent(Long toolContentId, boolean removeSessionData)
-	    throws SessionDataExistsException, ToolException {
+    public void removeToolContent(Long toolContentId) throws ToolException {
 	Scratchie scratchie = scratchieDao.getByContentId(toolContentId);
-	if (removeSessionData) {
-	    List list = scratchieSessionDao.getByContentId(toolContentId);
-	    Iterator iter = list.iterator();
-	    while (iter.hasNext()) {
-		ScratchieSession session = (ScratchieSession) iter.next();
-		scratchieSessionDao.delete(session);
+	if (scratchie == null) {
+	    ScratchieServiceImpl.log.warn("Can not remove the tool content as it does not exist, ID: " + toolContentId);
+	    return;
+	}
+
+	for (ScratchieSession session : scratchieSessionDao.getByContentId(toolContentId)) {
+	    List<NotebookEntry> entries = coreNotebookService.getEntry(session.getSessionId(),
+		    CoreNotebookConstants.NOTEBOOK_TOOL, ScratchieConstants.TOOL_SIGNATURE);
+	    for (NotebookEntry entry : entries) {
+		coreNotebookService.deleteEntry(entry);
 	    }
 	}
+
 	scratchieDao.delete(scratchie);
     }
 
+    @Override
     public void removeLearnerContent(Long toolContentId, Integer userId) throws ToolException {
-	if (log.isDebugEnabled()) {
-	    log.debug("Removing Scratchie content for user ID " + userId + " and toolContentId " + toolContentId);
+	if (ScratchieServiceImpl.log.isDebugEnabled()) {
+	    ScratchieServiceImpl.log
+		    .debug("Removing Scratchie content for user ID " + userId + " and toolContentId " + toolContentId);
 	}
 
 	List<ScratchieSession> sessions = scratchieSessionDao.getByContentId(toolContentId);
@@ -1801,7 +1803,7 @@ public class ScratchieServiceImpl implements IScratchieService, ToolContentManag
 		    scratchieDao.removeObject(NotebookEntry.class, entry.getUid());
 		}
 
-		if (session.getGroupLeader() != null && session.getGroupLeader().getUid().equals(user.getUid())) {
+		if ((session.getGroupLeader() != null) && session.getGroupLeader().getUid().equals(user.getUid())) {
 		    session.setGroupLeader(null);
 		}
 
@@ -2034,7 +2036,7 @@ public class ScratchieServiceImpl implements IScratchieService, ToolContentManag
 	    // set options
 	    Set<ScratchieAnswer> newAnswers = new LinkedHashSet<ScratchieAnswer>();
 
-	    JSONArray answersData = (JSONArray) questionData.getJSONArray(RestTags.ANSWERS);
+	    JSONArray answersData = questionData.getJSONArray(RestTags.ANSWERS);
 	    for (int j = 0; j < answersData.length(); j++) {
 		JSONObject answerData = (JSONObject) answersData.get(j);
 		ScratchieAnswer answer = new ScratchieAnswer();
